+++ src/test/java/de/alpharogroup/collections/list/ListExtensionsTest.java
@@ -830,6 +830,31 @@ public class ListExtensionsTest
 	}
 
 	/**
+	 * Test the method {@link ListExtensions#toArray(List)}
+	 */
+	@Test
+	public void testListToArray()
+	{
+		Integer[] actual;
+		Integer[] expected;
+		List<Integer> numbers;
+		// new scenario...
+		numbers = ListFactory.newArrayList(1,2,3);
+		actual = ListExtensions.toArray(numbers);
+		expected = ArrayFactory.newArray(1,2,3);
+		assertTrue(Arrays.equals(actual, expected));
+	}
+
+	/**
+	 * Test the method {@link ListExtensions#toArray(List)} that throws an <code>IllegalArgumentException</code>
+	 */
+	@Test(expectedExceptions = IllegalArgumentException.class)
+	public void testListToArrayWithException()
+	{
+		ListExtensions.toArray(ListFactory.newArrayList());
+	}
+
+	/**
 	 * Test the method {@link ListExtensions#toList(Enumeration)}
 	 */
 	@Test